Fixed Seeding Bug
There was a bug on shells which high curvature where the seeding would stop.  This was due to the new seed points being too close to the existing stress lines as the point was snapped to the mesh surface. The fix was to decrease the distance checking parameter for new seed points.